backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sse-starlette>=2.0.0
anthropic>=0.40.0
google-genai>=1.0.0
langgraph>=0.2.0
langchain-core>=0.3.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
